instance: Update certutil password handling
To quote other Puppet contributors:

This Misconception lies in the Code since the most early Days.

When adding one's own SSL bundle, the cert db is generated with a
password, but the password was never used by the 'exec' that changes the
trust settings. This results in 'certutil' trying to open a terminal to
prompt for the password, which fails, which crashes the Puppet run.

The short version is that the temporary password file must always be
generated and used, not just when generating self-signed certs.