instance: Use cert_name for server certificate
Currently, there is a bug in that supplying 'cert_name' into the ssl
hash will cause the Exec that changes the trust to operate on that name,
however the pk12 creation step uses the server fqdn for the name. The
result is the trust-modify Exec failing, crashing the Puppet run.